auto track

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/mechanisms/shooter/ServoTurret.java
@@ -7,7 +7,7 @@ import com.pedropathing.ivy.commands.Wait;
 import com.pedropathing.ivy.groups.Sequential;
 import com.qualcomm.robotcore.hardware.HardwareMap;
 
-import org.firstinspires.ftc.teamcode.utils.AllianceColor;
+import org.firstinspires.ftc.teamcode.utils.commands.AllianceColor;
 import org.firstinspires.ftc.teamcode.utils.Globals;
 import org.firstinspires.ftc.teamcode.utils.commands.Conditional;
 import org.firstinspires.ftc.teamcode.utils.hardware.HwServo;
@@ -17,10 +17,10 @@ import java.util.function.Function;
 @Config
 public class ServoTurret extends HwServo {
     //When calibrating blue
-    public static Function<Float, Float> turretPos = f -> Globals.allianceColor == AllianceColor.Red ? 2 * ServoTurret.REST - f : f;
+    public static Function<Double, Double> turretPos = f -> Globals.allianceColor == AllianceColor.Red ? 2 * ServoTurret.REST - f : f;
 
     //When calibrating red
-    public static Function<Float, Float> turretPosInv = f -> Globals.allianceColor == AllianceColor.Blue ? 2 * ServoTurret.REST - f : f;
+    public static Function<Double, Double> turretPosInv = f -> Globals.allianceColor == AllianceColor.Blue ? 2 * ServoTurret.REST - f : f;
 
     public static double LEFT_TICKS_LIMIT;
     public static double RIGHT_TICKS_LIMIT;
